Update main.py
小修改

+++ main.py
@@ -58,7 +58,7 @@ class DownloadMusic():
 
         SongFileName = re.sub(r'[\\/:"*?<>|]', '', SongName)
         if os.path.exists(rf'{self.DownloadPath}\{SongFileName}.mp3'):
-            print("歌曲已存在，跳过下载")
+            print("歌曲已存在,跳过下载")
             return True
 
         # 根据歌曲源下载歌曲
@@ -169,12 +169,16 @@ class DownloadMusic():
         try:
             res = requests.get(url, headers=self.header)
 
-            # 如果下载不到封面图片，使用默认封面避免最终文件损坏
+            # 如果下载不到封面图片,使用默认封面避免最终文件损坏
             if '.jpg' in Output and len(str(res.content)) >= 1:
-                print('下载封面失败，使用默认封面')
-                copyfile(rf'{self.RunPath}\NoImage.jpg',
-                         rf'{self.TempPath}\cover.jpg')
-                return True
+                if os.path.exists(rf'{self.RunPath}\NoImage.jpg'):
+                    print('下载封面失败,检测到NoImage.jpg,使用该文件作为封面')
+                    copyfile(rf'{self.RunPath}\NoImage.jpg',
+                             rf'{self.TempPath}\cover.jpg')
+                    return True
+                else:
+                    print('下载失败:下载封面失败\n放置NoImage.jpg以作为默认封面使用')
+                    return False
 
             # 如果下载到的是网页内容
             elif '<!DOCTYPE html>' in str(res.content):
@@ -234,7 +238,7 @@ def Main(file: str = 'listen1_backup.json', OutputPath: str = 'output') -> None:
     # Data[PlaylistID[Select]]
     SelectPlaylistData: dict = Data[PlaylistID[Select]]["tracks"]
     SelectPlaylistName: str = Data[PlaylistID[Select]]['info']['title']
-    print(f'正在准备下载 {SelectPlaylistName} 中的所有歌曲，共有{len(SelectPlaylistData)}首歌')
+    print(f'正在准备下载 {SelectPlaylistName} 中的所有歌曲,共有{len(SelectPlaylistData)}首歌')
     print('-'*50)
 
     # 创建对象
@@ -255,9 +259,9 @@ def Main(file: str = 'listen1_backup.json', OutputPath: str = 'output') -> None:
         # time.sleep(0.5)
 
     print("歌曲下载完毕")
-    print(f"共{len(SelectPlaylistData)}首歌，{Done}首下载成功，{Fail}首下载失败")
+    print(f"共{len(SelectPlaylistData)}首歌,{Done}首下载成功,{Fail}首下载失败")
     # 清理temp目录
-    # rmtree(D.TempPath)
+    rmtree(D.TempPath)
 
 
 if __name__ == '__main__':
